Google Gemini/Converse with Documents A Langchain & Google Gemini Pro App/Code/requirements.txt
streamlit
langchain
google-generativeai
python-dotenv
faiss-cpu
pypdf2
langchain_google_genai